requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.0
Flask==3.1.2
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==26.0
pillow==12.1.0
PyMySQL==1.1.0
cryptography==47.0.0
reportlab==4.4.9
requests==2.32.5
urllib3==2.6.3
Werkzeug==3.1.3
python-dotenv==1.0.1